tera things: render something with it
Not all my templates are supported by tera yet although they sort of
adhere to the ninja2 docs, so add templates2/ to contain some files for
testing and use them.

Also, having multiple levels of parents does not seem to be supported
yet in tera, so I'm using a local repo where it's fixed, with
.cargo/config containing its path.

Ugly code at the moment, work in progress.