ancient stuff for localized index page
This is also years old. Add a block to specify the index page's url in
the per-language base templates. The leaf pages don't show this, nor do
they actually show the tagline either; tera of this version doesn't
propagate the elements through the middle nodes but only from the leaf.
Probably a bug.