plan translations linkage
manual tracking of translations in the original would be manual and less
robust but more explicit and would need less magic in the core code, so
perhaps something like this?

Or detect a page type in the yaml and go build a magic linkage
automagically? What would it be called in the destination node then?