extract the document title from the rst source
This sort of surgical precision with types is a bit "interesting" with
no documentation; the learning curve is steep but not too tall. Maybe
there's a better way but it's undocumented. Next up to actually use the
title also in the translation mapping data. An extra pass over the pages
would render them first and collect this.