rudimentary group support
The metadata tags are collected together so that content items can be
grouped by anything:

"ok: true" (for now) becomes a group of non-draft pages
"language: en" becomes a group of pages that have *a language* set, not
a group of the keyword "en"
"category: [foo, bar]" similarly produces a group of pages that have a
category specified at all, not groups of pages in the foo and bar
categories.

This should be fine as long as the group can be filtered further down in
the templates.