implement lists of categories
The categories are only available via groups that contain pages but not
the metadata values (category names): the values already exist in the
pages, and they're often all identical (such as the groups "ok" or
"language") so it would be silly to copy them. Looking inside the
category arrays for the category names is site-specific, so this has to
be done in templates. Thus:

Read the category group's pages, find those with right language, reduce
to the page metadata's category value, flatten the list of category
lists (getting a lot of duplicates), then run that through unique and
finally sort the resulting list of category names.

Tera doesn't know about flattening arrays, so add a template filter for
it. This is general enough to consider upstreaming, but error handling
inside the flat_map iterators gets too complicated for tonight. Perhaps
this should be written as a plain simple loop to extend a vec. On the
other hand iterator wrestling is good rust practice.