prototype content cross ref ideas a bit
Assume that the metadata key "original" links to a page written in a
certain language and that pages in other languages are derivations of
them, and inject links back to the pages with an original link set. This
way the actual original page can display a link to other languages
without redundant metadata about those other languages. This "graph
edge" between the pages is undirected so it's best to write it only
once.

Of course, the original page does not need to actually be the first
version to be translated; the name just points to a consistent language
that serves as the "root" of a group of pages.