implement the Nuutti mode of translation matching
Multilingual pages can actually be cross-linked via the grouping feature
as Nuutti suggested (thanks). Add a demo. Have all language variants of
a single page specify a "translation_key" metadata with a value unique
to the page. The unique value can be anything; trivially it could be the
url of one of the language variants.

Not needing to implement the crossrefs just to support page translations
would be brilliant; less is more. However, both this grouping and the
orig crossrefs have their pros and cons, such as:

- Translation keys need to be exactly identical in each page variant, so
  typo bugs are possible. The orig links also need to be stated in all
  but one variant, but the referencing could verify that the destination
  page actually exists. Typo bugs could be mitigated by double checking
  that things change when a translated version has been written.
- The original link destination comes naturally from the implicitly
  defined url. The translation key needs to be invented somewhere, so
  either an arbitrary brain-based key allocator needs to be used each
  time producing possibly inconsistent results or the url has to be
  repeated in the original page metadata. Alternatively the rotuli core
  code could support a magical "url of this particular page" tag, but
  non-obvious magic features are evil.

Also, original vs. translations:

- The translations list in the orig page shows everything in a single
  place; the original links or translation keys would be scattered
  around in each page variant. But the translations list would need to
  redundantly specify the language of each translated variant which is
  fairly harmless but certainly not ideal design.

The crossrefs feature might be very useful for something else, too,
although that needs to be figured out before it pays off to implement.
Having the translation variant information "emerge" naturally from a
group does seem like the right thing to do even with some redundancy in
metadata and urls.