add prev and next functionality via filters
Add two filters, before_attr and after_attr, to find an element in a
list before or after an element with a specified attribute having some
certain value. Demo this with previous and next blog entries. Error
handling is still a bit WIP.

Prev and next by date or alphabetic order or anything is easily done by
sorting the list of pages first and then applying the prev or next
filter.