assume missing {before,after}_attr attrs falsy
Don't crash on missing attr but act as if the attr just wouldn't match.
This behaviour is consistent with the "undefined variables are
considered falsy" of template conditionals and allows more flexible
matching in things like looking up stuff based on metadata keys that may
not exist on all pages.

Returning an error would be also acceptable but would require the filter
chain to firts guarantee (with the "filter" filter) that the input
always has the attribute that is being matched.