experiment with blog dates and sorting
Add creation dates to the blog entries in one language and sort the
pages by that metadata key when listing them.

Fortunately Tera seems to use vec::sort_by_key() which is stable, so
sorting by title first and then by date produces consistent alphabetic
ordering when entries have only a date and when many have the same date.
Ideally all entries would also have an exact time with no collisions to
the level of seconds in practice, but that may not be always available.

serde-yaml / yaml-rust don't know about timestamp types so the creation
timestamps are strings. That works out very well for sorting when using
the iso8601 format that is well ordered in text form (except for
timezones, but those won't be needed in a while). Comparing timestamps,
doing arithmetic or retrieving individual parts of them (e.g., just the
year) would be cumbersome though. Someone should implement timestamps.

Sorting the page listings in the templates does not exactly align with
the design goal of being extremely fast because this means that the list
of entries would get sorted in every rendered page. This is a lot of
redundant work for a clean build (even if in practice sorting only 100s
of items wouldn't exactly be too slow), but for incremental changes only
one page or few pages would get changed at a time. Supporting sorted
entries of groups like this would not align with the design goal of
being generic and data-agnostic. Where would the sort key even be
specified? Perhaps have it in the template, and then cache the results?