add category listing to sample blog page template
List the category array from metadata if it exists, linking to the
category listing pages. This mechanism trusts that the category pages as
such exist for every category specified in the blog entries.

The comma-separated stuff is a bit clumsy like this, but it's written
only once, in the template.