retrieve the first body paragraph as a summary
Add a core rotuli feature similar to the doc title because parsing the
input markup is a core feature: find the content object after the title,
and if it's a paragraph, it is taken as the "summary" text. Sometimes
cutting it short to some number of sentences might be a good feature as
well, but here the content author is expected to write short enough
intro paragraphs.

Use this for the blog index page. Add some extra formatting in various
places to demonstrate how the document tree content changes.

The paragraphs per list item are an interesting change; apparently
rst2html4.py doesn't do that, but rst2html5.py does. See also
https://www.hawkeslearning.com/Accessibility/guides/html_content.html
"Almost all text (tables and headers being exceptions) will need to be
wrapped in p tags to display correctly. This includes inside li and dd
tags.", "For semantic purposes, content within the li element should be
wrapped in a p tag"