add mental note about title dilemma
The title is special and technically not part of the body text, so it
would be nice to be able to inject extra stuff after the title but
before the body. Not using the title at all in the rst content but
writing it in the metadata would be one option, but the top level title
produces a html section with the same id. This section might be useful
for some things.

The displayed metadata is left like this for now. Perhaps it could float
to some box to hide in a corner or a navigation panel because it's not
super useful anyway.