remove translations and some debug stuff
The core code doesn't need to support translations because they're
easily doable with metadata mangling in templates.

The direct references and automatic backlinks may get implemented later
if it proves to be superior to original-agnostic grouping (which is
superior to having to link both to the original in translated pages and
from the translations in the original page).

Cleanup also some minor debug things and notes.