abandon automatic crossrefs for now
For what the references were specifically needed, the groups can work
as-is. Translation groups are natural collections of page versions,
although that does add some redundancy in the metadata. That's the price
for a minimal core.

Add a bunch of historical documentation near the translation link macro
to serve as a specification if these refs become useful in the future.