drop source/translations mapping in sample source
Use the more general translation_key feature and delete the original
links and translations maps that require care with bidirectional refs.

Add a language listing in the base-en template to demo the frontpage
written in dog language. For brevity, there's no link to it, but its
existence is displayed. It's accessible by writing the url manually.