make url stuff more consistent
Use the display_url() method that appends the slash for things that
shall become directories. The httpd usually would redirect to the right
url from links that don't have the slash, but be explicit about it.

Move one lang version of the category listing to an index.rst in the
same directory to demo the difference that's not a difference in the
rendered urls.