add take_until_attr filter
The before_attr and after_attr filters are nice, but they're too
specific in that the same functionality can be implemented fairly easily
with more general filters that can do more: an object just before X in a
sequence is the last one in a slice that ends just before X, and
sometimes having the whole slice can be useful. Add take_until_attr()
and reimplement the blog previous and next link demos with it. The next
link doesn't need a "take_after_attr" filter: just reverse the input,
use take_until_attr.

If this copies the whole list of pages instead of making a slice
internally, perf could become an issue with massive lists. In such cases
the before_attr and after_attr could become useful, but that'll need a
benchmark later.