start pondering page date hierarchy
Add a year template for listing blog entries from one year. Source
documents for these need to exist in the filesystem. It's not such a big
deal because good filesystems support symbolic links.

Generating these on the fly would break the input/output hierarchy
consistency, even if the engine could somehow link to one master copy
for all generated pages. With such design, the metadata would need a
magic key, such as generate_index_page: /\d{4}/\d{2}/\d{2}/ to compare
against for each directory that's generated in the output.

Now there's yet another dilemma. The url hierarchy looks like a
directory hierarchy, but it only works well for just one language.

The index pages for the date hierarchy:

- /2020/7/21/
- /2020/7/
- /2020/

for pages in different languages in both:

- /2020/7/21/helloworld/
- /2020/7/21/päivää/

is troublesome. And having the language in the url is not ideal:

- /en/2020/7/21/helloworld/
- /2020/7/21/päivää/

the following is not acceptable because the current blog would break:

- /en/2020/7/21/helloworld/
- /fi/2020/7/21/päivää/

This certainly would not work:

- /2020/7/21/en/helloworld/
- /2020/7/21/päivää/

because only the full date would be correct:

- /2020/7/21/en/helloworld/

and the month and the year would still be ambiguous:

- /2020/7/
- /2020/

A mixed language page for the listings is possible but not ideal.