read markup and index strings from user
Add cli opts for the markup language which will also be the file suffix
and for the directory index filename which is used to decide whether a
file produces a directory or exists for the page only. The markup
language gets parsed to and from an enum type; for now just rst exists.

Perhaps the index feature should go because duplicates are possible (and
not yet detected). Just foo.rst and foo/ would work, although that would
separate blog entries and their files from each other; not sure if
that's good or bad.