don't repeat metadata in context
This is already available in page.meta and adding it again for
convenience has a cost, so don't do it. Accessing it via page.meta is
more explicit, less magical, and thus better.

Btw, whether to insert the site-common data (like now) or to make a
prebuilt context to clone and extend from for each page needs to be
benched. Either the data is reserialized or it's cloned; the latter
seems to require more allocations at the moment.