copy plain files in source dir
The files identified by the specified markup language are rendered; the
rest needs to exist as well, so for now just copy those files around.
This could be more efficient with other tools (e.g., rsync) and
filesystem features (soft/hard links) but copying by rotuli eases
development. The --render-only flag could be used to suppress copies.

Add also the css file that's referred to by the sample html to demo.

The repetitive create_dir_all should be benched or rather separated from
the per-file process and done beforehand for all dirs as we know all
dirs that would be accessed (both for rendering and copying).