list orphan blog pages via api
It can be useful to check that all blog entries belong to some category.
Use the Rust side to implement this; having a page and a custom template
would also be possible, but it's not always useful to have a web page
visible for everyone when only the blog maintainer needs to see the
list.

This is also the first example of how to use the core code as a library
for something else than rendering the site.