add an image to try that aspect of rust-rst
This renders well enough, but rust-rst seems to have some trouble with
parsing (and ignoring) whitespace around the alt text in the image
directive. More than one option (e.g., alt and target) are apparently
not parsed properly yet either. That will pose a problem with thumbnails
and links to originals.