add by-value lookup to group page listings
A very common use case for page groups is to list some pages in the
group by a value of that group key. The group name is a metadata key and
the group contains all pages with that key, regardless of value. Lookup
by value is more readable than filtering the pages that have the key.

So in template code, instead of:

    site.groups["language"].pages | filter(attribute="meta.language", value="en") | ...

we have:

    site.groups["language"].pages_by_str_value["en"] | ...

This is more readable but the one fewer filter step does not noticeably
speed up my blog (WIP) of ~230 posts. Might be worthwhile to look into
the custom function support in Tera.